extracting cache coordinator
